pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "typst-mpl-backend"
description = "Typst backend for matplotlib (Python visualization library)."
license = {text = "MIT"}
authors = [
    {name = "Daniel Bershatsky", email = "daniel.bershatsky@gmail.com"},
]
readme = {file = "README.md", content-type = "text/markdown"}
version = "0.1.0"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Other Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Text Processing",
    "Topic :: Text Processing :: Markup",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = ["matplotlib", "numpy"]
requires-python = ">=3.11,<4"

[project.optional-dependencies]
dev = ["isort", "pytest>=8", "ruff"]
test = ["pillow", "numpy"]

[project.urls]
Homepage = "https://github.com/daskol/typst-mpl-backend"
Repository = "https://github.com/daskol/typst-mpl-backend.git"

[tool.isort]
combine_as_imports = true
known_first_party = ["mpl_typst"]
multi_line_output = 4

[tool.mypy]
no_implicit_optional = true

show_column_numbers = true
show_error_codes = true
show_error_context = false

warn_unused_ignores = true

plugins = ["numpy.typing.mypy_plugin"]

[tool.pytest.ini_options]
minversion = "8.2"
addopts = "-ra -q -m 'not slow'"
testpaths = ["mpl_typst"]
markers = [
    "slow: slow tests (deselect with '-m \"not slow\"')",
]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
indent-width = 4
line-length = 79
extend-exclude = ["*_pb2.py"]


[tool.ruff.format]
# Prefer single quotes over double quotes
quote-style = "single"

[tool.ruff.lint]
select = ["E", "F", "NPY", "PYI", "PTH", "Q"]
ignore = ["PTH123"]
preview = true

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"

[tool.ruff.lint.pycodestyle]
max-doc-length = 79
max-line-length = 79

[tool.setuptools]
zip-safe = true

[tool.setuptools.packages.find]
include = ["mpl_typst*"]

[tool.setuptools.package-data]
mpl_typst = ["*.typ"]
